refactor: handshake sink, stream and builder
